requirements.txt
# Version Information
# Last Updated: 2025
# Author: Keith Satuku
# License: MIT 

# Core dependencies
numpy>=1.21.0,<2.0.0
pandas>=1.3.0,<2.0.0
pydantic>=2.7.4,<3.0.0
PyYAML>=6.0.0

# ML and Embeddings
torch>=2.1.0  # Base PyTorch
sentence-transformers==2.5.1
transformers>=4.33.0
huggingface-hub>=0.25.0
tokenizers>=0.15.2
flagembedding>=1.2.11
datasets>=2.18.0
peft>=0.14.0
faiss-cpu>=1.7.4

# Document Processing
python-magic>=0.4.27
python-docx>=0.8.11
pdf2image>=1.16.0
Pillow>=8.3.0
pytesseract>=0.3.8

# Audio Processing (Optional)
librosa>=0.8.1
soundfile>=0.10.3
speechbrain>=0.5.12

# Vector Store
qdrant-client>=1.7.0

# Utilities
python-dotenv>=0.19.0
tqdm>=4.65.0
chardet>=5.0.0
langdetect>=1.0.9

# API
fastapi>=0.68.0,<1.0.0
python-multipart>=0.0.5
typing-extensions>=3.10.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.18.0
pytest-cov>=3.0.0
pytest-benchmark>=3.4.1

# Development
black>=23.0.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.5.0

# Embedding and ML
pydantic>=2.7.4,<3.0.0
pytest>=7.0.0
PyYAML>=6.0.0

# Audio processing
openai-whisper>=1.0.0
speechbrain>=0.5.12


# Utilities
python-dotenv>=0.19.0
tqdm>=4.65.0

# Image Processing
opencv-python>=4.8.0
Pillow>=8.3.0
pytesseract>=0.3.8
rdkit>=2023.3.1

# Machine Learning
scikit-image>=0.21.0
scikit-learn>=0.24.2

# Document Processing & OCR
beautifulsoup4>=4.12.0
lxml>=4.9.0

# NLP and Vision
nltk>=3.6.0
spacy>=3.7.0
tiktoken>=0.3.0
keybert>=0.7.0

# Testing & Development
black>=23.0.0

# Document Processing
ftfy>=6.1.1

# Visualization and Analysis
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# Caching and Database
redis>=4.0.0

# Development Tools
isort>=5.12.0
flake8>=6.1.0
mypy>=1.5.0

# Documentation
sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0

# Core ML and data processing
rank-bm25>=0.2.2

# Dependency injection and configuration
dependency-injector>=4.41.0

# Vector operations and embeddings
datrie>=0.8.2
editdistance-s>=1.0.0
hanziconv>=0.3.2
json-repair>=0.8.0
roman-numbers>=0.3.0
word2number>=1.1
cn2an>=0.5.22
grpcio>=1.59.0
protobuf>=4.24.0

# Document processing
markdown-to-json>=2.1.0
latex2sympy2>=1.0.0
ranx>=0.3.8
umap-learn>=0.5.4
pynvml>=11.5.0

# Image and diagram processing
ray>=2.9.0
PyMuPDF>=1.18.0
whisper-openai>=1.0.0

# API and Web
fastapi>=0.68.0,<1.0.0
python-multipart>=0.0.5
typing-extensions>=3.10.0

# Audio processing
torchaudio>=2.1.0
torchvision>=0.10.0

# Validation
validators>=0.20.0

# Testing
pytest-xdist>=2.5.0
hypothesis>=6.0.0
faker>=8.0.0
aiohttp>=3.8.0

#Advanced ML
graspologic>=3.3.0
umap-learn>=0.5.4

# Documentation
sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0 